move remove edsc and move gifstream into modules #3647
