+++ .eslintignore
@@ -10,6 +10,5 @@ web/ext/*
 web/dist/*
 web/build/*
 test/*
-web/js/edsc/*
 web/js/lib/*
