+++ web/js/containers/gif.js
@@ -7,7 +7,7 @@ import { debounce as lodashDebounce, round as lodashRound } from 'lodash';
 import {
   Progress, Modal, ModalBody, ModalHeader,
 } from 'reactstrap';
-import GifStream from '../lib/gifstream';
+import GifStream from '../modules/animation/gifstream';
 import GifPanel from '../components/animation-widget/gif-panel';
 import util from '../util/util';
